Move order logic into a-scene.js and introduce callComponentBehaviors method
